+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["src/"]
+packages = ["PCA/"]
 
 [project]
 name = "PCA"
