+++ tests/unit_tests.py
@@ -3,17 +3,17 @@
 import pytest
 import numpy as np
 
-from src.PCA import *
+from PCA.PCA import *
 from sklearn.datasets import load_iris, make_low_rank_matrix
 
 # global_seed ?
 
-
+# Define different input datasets
 @pytest.fixture(scope="module")
 def random_data():
     """Use random Matrix"""
     # rng = np.random.RandomState(0)
-    np.random.seed(0)
+    np.random.seed(0) # use a seed for reproducibility
     mat = np.random.rand(100, 1000)
     # rng(100, 1000)  # random size ?
 
@@ -22,15 +22,20 @@ def random_data():
 
 @pytest.fixture(scope="module")
 def iris():
-    """Use iris dataset"""
+    """Use iris dataset - 
+       Classic dataset used to demo PCA or classification problems.
+    """
     iris_dataset = load_iris()
     return iris_dataset.data
 
 
 @pytest.mark.parametrize("data_fixture", ["random_data", "iris"])
 def test_data_centered(data_fixture, request):
-    """Check that remove_mean returns a matrix of mean 0"""
+    """Check that remove_mean returns a matrix of mean 0
+       Fixture is used to run same test on several input datasets.
+    """
     mat = request.getfixturevalue(data_fixture)
+
     # data_fixture
     centered_mat = substract_mean(mat)
     assert np.any(
@@ -55,8 +60,12 @@ def check_covariance_identity():
 # @pytest.mark.parametrize("data_fixture", ["random_data", "iris"])
 def test_covariance_correlated():
     """Covariance matrix values of rank 1 matrix are norm of column * factor"""
+
+    # generate random vector
     np.random.seed(0)
     rand_vec = np.random.rand(100)
+
+    # generate rank 1 matrix by stacking this random vector, multiplied by some coefficients
     rank1_matrix = np.transpose(np.stack([(u + 1) * rand_vec for u in range(10)]))
     cov_mat = get_covariance_matrix(rank1_matrix)
     for i in range(10):
@@ -69,8 +78,6 @@ def test_covariance_correlated():
 @pytest.mark.parametrize("data_fixture", ["random_data", "iris"])
 def test_covariance_symmetric(data_fixture, request):
     """ Covariance Matrix should alway be symmetric"""
-    # rng = np.random.RandomState(0)
-    # mat = rng(100, 1000)  # random size ?
     mat = request.getfixturevalue(data_fixture)
     cov_mat = get_covariance_matrix(mat)
     assert np.all(cov_mat == np.transpose(cov_mat))
